feat: complete real cryptographic operations implementation
- Remove placeholder proof type from ZKP verification
- Fix Schnorr proof verification logic (proper commitment handling)
- Add comprehensive test suite for ZKP (4 tests)
- Verify Paillier HE implementation is complete and functional
- Add comprehensive test suite for HE (7 tests)
- Create SecureAggregator::with_paillier_encryption() factory method
- Add integration tests for cryptographic operations
- Document complete implementation status

All cryptographic operations now use real implementations:
- Zero-Knowledge Proofs: Real Schnorr proof-of-knowledge
- Homomorphic Encryption: Real Paillier cryptosystem
- No placeholders remain in production code paths

+++ iterations/v3/agent-orchestration/src/planning/data_processing_adapter.rs
@@ -477,7 +477,7 @@ impl DataProcessingService for DataProcessingServiceAdapter {
             successful: stats.total_processed.saturating_sub((stats.total_processed as f64 * stats.error_rate) as u64),
             failed: (stats.total_processed as f64 * stats.error_rate) as u64,
             average_processing_time_ms: stats.avg_processing_time_ms,
-            queue_size: stats.queue_depth,
+            queue_size: stats.queue_depth as u64,
             success_rate,
         })
     }
